+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+## [1.2.0] - 2020-11-21
+
 ### Added
 
 - Serialization of local components in Blackboard Container.
@@ -32,6 +34,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Unity editor support.
 - Tests.
 
-[unreleased]: https://github.com/ZorPastaman/Simple-Blackboard/compare/v1.1.0...HEAD
+[unreleased]: https://github.com/ZorPastaman/Simple-Blackboard/compare/v1.2.0...HEAD
+[1.2.0]: https://github.com/ZorPastaman/Simple-Blackboard/releases/tag/v1.2.0
 [1.1.0]: https://github.com/ZorPastaman/Simple-Blackboard/releases/tag/v1.1.0
 [1.0.0]: https://github.com/ZorPastaman/Simple-Blackboard/releases/tag/v1.0.0